backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
rank-bm25==0.2.2
requests==2.32.3
